requirements.txt
adafruit-ampy==1.1.0
awscrt==0.24.1
awsiotsdk==1.22.2
bitarray==3.2.0
bitstring==4.3.1
cffi==1.17.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
ecdsa==0.19.1
esptool==4.8.1
intelhex==2.3.0
pycparser==2.22
pyserial==3.5
python-dotenv==1.0.1
PyYAML==6.0.2
reedsolo==1.7.0
six==1.17.0
